==> codes/Step 2 modelling surprise SVD.ipynb <==
## importing the libraries
import pandas as pd
import numpy as np
from surprise import Dataset, Reader, SVD
from sklearn.decomposition import PCA
from sklearn.compose import make_column_transformer
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.tree import DecisionTreeClassifier
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score


## setting the maximum columns and rows
import os
pd.set_option('display.max_columns', 500)
pd.set_option('display.max_rows', 500)

## removing warnings
import warnings
warnings.simplefilter("ignore", UserWarning)
warnings.simplefilter(action='ignore', category=FutureWarning)
pd.options.mode.chained_assignment = None

## importing datasets
master_data = pd.read_csv(r"../data/processed_data/master_data_for_modelling.csv")
data = master_data.iloc[:,2:]

## converting categorical columns into strings
categorical_cols = ['flag_ip','Gender','Race', 'RenalDiseaseIndicator','ChronicCond_Alzheimer','ChronicCond_Heartfailure','ChronicCond_KidneyDisease','ChronicCond_Cancer',
'ChronicCond_ObstrPulmonary','ChronicCond_Depression','ChronicCond_Diabetes','ChronicCond_IschemicHeart','ChronicCond_Osteoporasis',
'ChronicCond_rheumatoidarthritis','ChronicCond_stroke','Aged','community_institutional','Disability_condn','Benefits',
'medicaid_flag','Disease_intraction_DIABETES_HF','Disease_intraction_HF_CHR_LUNG','Disease_intraction_HF_KIDNEY','Disease_intraction_CHR_LUNG_CARD_RESP_FAIL',
'Disease_intraction_HF_HCC238','Disease_intraction_gSubUseDisorder_gPsych_','DISABLED_HF','DISABLED_ULCER_',
'DISABLED_CANCER','DISABLED_NEURO','DISABLED_CHR_LUNG','disability']
data[categorical_cols] = data[categorical_cols].astype(str)

## saclling and onehot encoding of data

# Define the column transformer
ct = make_column_transformer(
    (StandardScaler(), data.select_dtypes(include=['int', 'float']).columns.tolist()),
    (OneHotEncoder(drop = "first"), data.select_dtypes(include=['object']).columns.tolist()))

# Define the pipeline
pipeline = make_pipeline(ct)

# Transform the data
transformed_data = pipeline.fit_transform(data)

# Get the names of the encoded columns
encoded_columns = pipeline.named_steps['columntransformer'].transformers_[1][1].get_feature_names(data.select_dtypes(include=['object']).columns.tolist())

# Combine the column names
column_names = data.select_dtypes(include=['int', 'float']).columns.tolist() + encoded_columns.tolist()

# Convert the transformed data to a DataFrame with column names
transformed_data_df = pd.DataFrame(transformed_data, columns=column_names)


## missing values imputation with median
transformed_data_df = transformed_data_df.fillna(transformed_data_df.median())

## Decision Tree

## finding top features based on decision tree
X = transformed_data_df.drop(columns="other_disease_mean_hcc_factor")
y = master_data['HCC']

# Train the decision tree model
model = DecisionTreeClassifier()
model.fit(X, y)

# Get the feature importances
importances = model.feature_importances_

# Create a new DataFrame with the feature names and importances
feature_importances = pd.DataFrame({'feature': X.columns, 'importance': importances})

# Sort the features by importance (descending order)
feature_importances = feature_importances.sort_values('importance', ascending=False).reset_index()
top_features = feature_importances['feature'][:19]

## random forest model

##creating a classification model on topfeatures to see its performance
X = transformed_data_df[top_features]
y = master_data['HCC']

# Split the data into train and test sets (70% training, 30% testing)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

# Train a random forest model with 100 trees
model = RandomForestClassifier(n_estimators=100)
model.fit(X_train, y_train)

# Use the trained model to predict the labels of the test set
y_pred = model.predict(X_test)

# Calculate the accuracy, precision, recall, and F1-score
accuracy = accuracy_score(y_test, y_pred)
precision = precision_score(y_test, y_pred,average="macro")
recall = recall_score(y_test, y_pred,average="macro")
f1 = f1_score(y_test, y_pred,average="macro")

# Print the results
print('Accuracy:', accuracy*100)
print('Precision:', precision*100)
print('Recall:', recall*100)
print('F1-score:', f1*100)

matrix with all features-

- Accuracy: 31.020968759367046
- Precision: 28.64776065297239
- Recall: 08.417712716824272
- F1-score: 10.492380559764748



## PCA

# Apply PCA to combine the variables into a single factor
pca = PCA(n_components = 0.8)
factors_PCA = pca.fit_transform(transformed_data_df[top_features])

pca.explained_variance_ratio_

## LDA

# Apply LDA to combine the variables into a single factor
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
clf = LinearDiscriminantAnalysis(n_components = 1)
factors_LDA = clf.fit_transform(transformed_data_df[top_features], master_data['HCC'])

## checking variance captured by top LDA feature created using Top 19 raw features
clf.explained_variance_ratio_

## MDA

## using MDA to get the variance capture
import prince
# Fit and transform the MCA
mca = prince.MCA(n_components=1, n_iter=3, copy=True, check_input=True, engine='sklearn')
mca.fit(transformed_data_df[encoded_columns])
# mca_transformed = mca.transform(transformed_data_df[encoded_columns])

# Get the explained variance ratio
mca.eigenvalues_summary

## FAMD

## using FAMD to create a feature

import prince
# Fit and transform the MCA
FAMD = prince.FAMD(n_components=1, n_iter=3, copy=True, check_input=True, engine='sklearn')
FAMD.fit(data.head(10000).dropna())
FAMD_transformed = mca.transform(data.dropna())

# Get the explained variance ratio
FAMD.eigenvalues_summary

## surprise SVD algorithm

# Add the factors to the data
master_data['factors'] = factors_LDA

## selecting data in a formate which Surprise library wants
recommender_data = master_data[['BeneID','HCC','factors']].drop_duplicates(subset=["BeneID","HCC"], keep = "first")

# Define the Reader
reader = Reader(rating_scale=(recommender_data["factors"].min(), recommender_data["factors"].max()))

# Load the data into Surprise's Dataset object
recommender_data_surprise = Dataset.load_from_df(recommender_data[['BeneID','HCC','factors']], reader)

# Define the SVD algorithm
algo = SVD()

# Train the algorithm on the data
trainset = recommender_data_surprise.build_full_trainset()
algo.fit(trainset)

# Make recommendations
user_id = 'BENE100178'
num_recommendations = 10
recommendations = []
for item_id in recommender_data_surprise.df['HCC'].unique():
    rating = algo.predict(user_id, item_id).est
    recommendations.append((item_id, rating))
recommendations.sort(key=lambda x: x[1], reverse=True)
top_recommendations = recommendations[:num_recommendations]
for recommendation in top_recommendations:
    print(recommendation)

master_data[master_data["BeneID"] == "BENE100178"]

Since the model is not performing well so will not use this model for recommendations

Possible reasons of inaccuracy-

1. we are using only 19 features to find the 1 master feature.
2. even this master feature is only able to capture 74% variablity from 19 features whearas total features are 81.
