==> codes/Step 4 modelling cosine similairty model based.ipynb <==
## importing the libraries
import pandas as pd
import numpy as np
import os
from sklearn.compose import make_column_transformer
from sklearn.pipeline import make_pipeline
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.metrics.pairwise import cosine_similarity
from scipy.sparse import csr_matrix
from scipy.sparse.linalg import svds

## setting display to show maximum 500 rows and columns
pd.set_option('display.max_columns', 500)
pd.set_option('display.max_rows', 500)

## ignoring the warnings
import warnings
warnings.simplefilter("ignore", UserWarning)
warnings.simplefilter(action='ignore', category=FutureWarning)
pd.options.mode.chained_assignment = None

## Load the data
master_data = pd.read_csv(r"../data/processed_data/master_data_for_modelling.csv")
df_features = pd.read_excel(r"C:../data/processed_data/all_features_for_modelling.xlsx")
data = master_data.iloc[:,2:]

master_data.head()

## defining and converting columns into categorical columns
categorical_cols = ['flag_ip','Gender','Race', 'RenalDiseaseIndicator','ChronicCond_Alzheimer','ChronicCond_Heartfailure','ChronicCond_KidneyDisease','ChronicCond_Cancer',
'ChronicCond_ObstrPulmonary','ChronicCond_Depression','ChronicCond_Diabetes','ChronicCond_IschemicHeart','ChronicCond_Osteoporasis',
'ChronicCond_rheumatoidarthritis','ChronicCond_stroke','Aged','community_institutional','Disability_condn','Benefits',
'medicaid_flag','Disease_intraction_DIABETES_HF','Disease_intraction_HF_CHR_LUNG','Disease_intraction_HF_KIDNEY','Disease_intraction_CHR_LUNG_CARD_RESP_FAIL',
'Disease_intraction_HF_HCC238','DISABLED_HF','DISABLED_ULCER_','DISABLED_CANCER','DISABLED_CHR_LUNG','disability']
data[categorical_cols] = data[categorical_cols].astype(str)

## selecting improtant features from all_features_for_modelling excel file
selected_features = list(df_features[df_features["selection_for_cosinesimilarity"]==1]["Features"])
data = data[selected_features]

## One-hot encoding and standard scalling

# Define the column transformer
ct = make_column_transformer(
    (StandardScaler(), data.select_dtypes(include=['int', 'float']).columns.tolist()),
    (OneHotEncoder(drop = "first"), data.select_dtypes(include=['object']).columns.tolist()))

# Define the pipeline
pipeline = make_pipeline(ct)

# Transform the data
transformed_data = pipeline.fit_transform(data)

# Get the names of the encoded columns
encoded_columns = pipeline.named_steps['columntransformer'].transformers_[1][1].get_feature_names(data.select_dtypes(include=['object']).columns.tolist())

# Combine the column names
column_names = data.select_dtypes(include=['int', 'float']).columns.tolist() + encoded_columns.tolist()

# Convert the transformed data to a DataFrame with column names
transformed_data_df = pd.DataFrame(transformed_data, columns=column_names)

## filling the missing values with median
transformed_data_df = transformed_data_df.fillna(transformed_data_df.median())

## adding Beneficiary Id in transformed data
transformed_data_df = pd.concat([master_data[['BeneID']] , transformed_data_df], axis = 1)

## checking head of transformed data
transformed_data_df.head()

## cosine similairty

#### user_user simialrity

## selecting train and test samples

## groupping data at beneficiary level and finding the median of each value at member level.
user_vars = transformed_data_df.groupby('BeneID').median()
df_user_vars_train = user_vars.iloc[:,:]
df_user_vars_test = user_vars.iloc[:,:]

## defining the class of recommendation model
class CosineRecommendation:

    def __init__(self, rank=50):
        self.rank = rank
        self.U = None
        self.S = None
        self.VT = None
        self.mape = None
        self.labels = None
 
    def fit(self, df):
        
        # storing index in labels
        self.labels = df.index

        # Perform matrix factorization using SVD with rank k
        self.U, self.S, self.VT = svds(csr_matrix(df), k=self.rank)
        
    def SVD_conversion_MAPE(self,df):

        # converting df into sparsh matrix
        df_spars_actual = csr_matrix(df).toarray()
        
        # Perform matrix factorization using SVD with rank k
        self.U, self.S, self.VT = svds(csr_matrix(df), k=self.rank)
        
        # Reconstruct the cosine similarity matrix using the low-rank approximation
        df_sparse_low_rank = np.dot(self.U, np.dot(np.diag(self.S), self.VT))
        
        # calculate mean absolute percentage error
        numerator = np.sum(np.abs(df_spars_actual - df_sparse_low_rank))
        denominator = np.sum(df_spars_actual)
        if denominator != 0:
            mape = 100 / denominator * numerator
        else:
            mape = np.nan

    def predict(self, new_data, top_recommendations = 5):
        
        '''
            new_data: panadas DataFrame which needs to be predcited
            top_recommendations :if top_recommendation is in between 0 to 1 then it will take the recommendation,
                                whoes simialrity is greater than top_recommednation.
                                if top_recommednation is greater than 1 then it will take the top_recommended recommendations
                                bydefault it will take top 5 recommendation
                                
                                default = 5
        '''
        
        # Reconstruct the cosine similarity matrix using the low-rank approximation
        df_sparse_low_rank = np.dot(self.U, np.dot(np.diag(self.S), self.VT))
        
        # Compute cosine similarity between the new datapoint and all the other datapoints in the training data
        cosine_sim_predicted_sparse = csr_matrix(cosine_similarity(df_sparse_low_rank, new_data)).toarray()

        # Clip the cosine similarity matrix to range between -1 and 1
        cosine_sim_predicted_sparse = np.clip(cosine_sim_predicted_sparse, -1, 1)

        # add labels in the data
        cosine_sim_predicted_pandas = pd.DataFrame(cosine_sim_predicted_sparse)
        
        #creating final recommendations
        final_recommendation = pd.DataFrame(columns = ['Target_Beneficiary', 'Similar_Beneficiary', 'Similarity_score'])
        
        for i in range(cosine_sim_predicted_pandas.shape[1]):
            
            if (max(cosine_sim_predicted_pandas.iloc[:,1]) < 0.8) & (top_recommendations < 1):
                continue
            else:
                pass
                
            temp_df = cosine_sim_predicted_pandas.iloc[:,[i]].rename(columns={i:"Similarity_score"})
            temp_df = pd.concat([pd.DataFrame(self.labels),temp_df],axis=1)
            temp_df = temp_df.sort_values("Similarity_score",ascending=False)
            temp_df = temp_df.reset_index(drop=True).rename(columns={"BeneID":"Similar_Beneficiary"})
            temp_df["Target_Beneficiary"] = new_data.index[i]
            temp_df = temp_df[["Target_Beneficiary","Similar_Beneficiary","Similarity_score"]]
            temp_df = temp_df[temp_df["Target_Beneficiary"]!=temp_df["Similar_Beneficiary"]]
            
            if (top_recommendations > 0) & (top_recommendations < 1):
                temp_df = temp_df[(temp_df["Similarity_score"].astype(float)) > top_recommendations]
                
            elif top_recommendations != 5:
                temp_df = temp_df.head(top_recommendations)
            
            else:
                temp_df = temp_df.head()

            ## appending temp_df in final_recommendations
            final_recommendation = final_recommendation.append(temp_df, ignore_index = True)

        return final_recommendation


%%time
# Create an instance of the CosineRecommendation class with rank k
recommendation_model = CosineRecommendation(rank=45)

## checking the MAPE of SVD factorization
recommendation_model.SVD_conversion_MAPE(df_user_vars_train.head(10000))

# Fit the model on the input dataframe
recommendation_model.fit(df_user_vars_train)

%%time
# defining the top recommendations
top_recommendation = 0.8  ## select all simialr beneficiaries whoes similairty score is greater than 0.8
number_of_test_split = 500  ## split the test data into 500 sub datasets. it will reduce the memroy usage

# splitting the test set into 
splitted_datasets = np.array_split(df_user_vars_test, number_of_test_split, axis=0)

# creating empty pataframe if all recommendation
all_recommendation = pd.DataFrame(columns = ['Target_Beneficiary', 'Similar_Beneficiary', 'Similarity_score'])

# finding predictions of all spillted test sets
for i in range(len(splitted_datasets)):
    recommendations = (recommendation_model.predict(splitted_datasets[i], top_recommendations = top_recommendation))
    all_recommendation = all_recommendation.append(recommendations, ignore_index = True)

## creating target_HCC and similar_HCC columns

df_BeneID_HCC_list = master_data[["BeneID","HCC"]].groupby("BeneID").agg(lambda x: set(x.to_list())).reset_index()
all_recommendation = all_recommendation.merge(df_BeneID_HCC_list, how = "left", right_on="BeneID", left_on = "Target_Beneficiary").drop(columns=["BeneID"]).rename(columns = {"HCC": "Target_HCC"})
all_recommendation = all_recommendation.merge(df_BeneID_HCC_list, how = "left", right_on="BeneID", left_on = "Similar_Beneficiary").drop(columns=["BeneID"]).rename(columns = {"HCC": "Simialr_HCC"})

## finding common HCC between target hcc and simialr hcc
all_recommendation['common_HCC'] = [x[0].intersection(x[1]) for x in zip(all_recommendation['Target_HCC'], all_recommendation['Simialr_HCC'])]
## finding common HCC counts
all_recommendation['common_HCC_count'] = [len(x) for x in all_recommendation['common_HCC']]

## sorting the recommendations based of firstly common_hcc_count and secondly similairty score
all_recommendation.sort_values(["Target_Beneficiary","common_HCC_count","Similarity_score"],ascending=[True, False,False],inplace=True)
all_recommendation = all_recommendation.groupby('Target_Beneficiary').apply(lambda x: x.head())
all_recommendation = all_recommendation.drop(columns=["Target_Beneficiary"]).reset_index().drop(columns=["level_1","common_HCC_count"])

## creating individual recommendations by individual similar service for target services and converting it into a list
all_recommendation["individual_recommended_HCC"] = all_recommendation["Simialr_HCC"] - all_recommendation["Target_HCC"]

## combining individual recommended HCC's at target beneficiary level and removing the duplicates.
## putting the HCC's by frequency counts
from collections import OrderedDict
df_combined_recommedned = all_recommendation[["Target_Beneficiary","individual_recommended_HCC"]].explode(
                                                column = "individual_recommended_HCC").dropna().groupby(
                                                "Target_Beneficiary").agg(lambda x: x.to_list()).rename(
                                                    columns={"individual_recommended_HCC":"combined_recommedned_HCC"}).reset_index()
# Function to sort and remove duplicates from the list
def sort_and_remove_duplicates(lst):
    lst = sorted(lst, key=lst.count, reverse=True)
    return list(OrderedDict.fromkeys(lst))

# Apply the function to the column containing lists
df_combined_recommedned['combined_recommedned_HCC'] = df_combined_recommedned['combined_recommedned_HCC'].apply(sort_and_remove_duplicates)
all_recommendation = all_recommendation.merge(df_combined_recommedned,how='left',on='Target_Beneficiary')
all_recommendation = all_recommendation.dropna(subset='combined_recommedned_HCC').reset_index(drop=True)

# Define a lambda function to slice the lists
trimmed_recommendations = 5
trim_list = lambda x: x[:trimmed_recommendations]

# Apply the lambda function to the column of lists
all_recommendation['combined_recommedned_HCC_trimmed'] = all_recommendation['combined_recommedned_HCC'].apply(trim_list)

## merging beneficiary level data on Target_Beneficiary to calculate RAF score
RAF_data_beneficiary_level = master_data[['BeneID', 'Gender', 'Age', 'Aged', 'community_institutional',
                                         'Disability_condn', 'disability',
                                           'Benefits', 'medicaid_flag', 'Disease_intraction_DIABETES_HF',
                                           'Disease_intraction_HF_CHR_LUNG', 'Disease_intraction_HF_KIDNEY',
                                           'Disease_intraction_CHR_LUNG_CARD_RESP_FAIL',
                                           'Disease_intraction_HF_HCC238', 'DISABLED_HF', 'DISABLED_ULCER_',
                                           'DISABLED_CANCER', 'DISABLED_CHR_LUNG']].drop_duplicates(subset=["BeneID"]
                                                                                                   ).reset_index(drop=True)
all_recommendation = pd.merge(all_recommendation,RAF_data_beneficiary_level, how='left',left_on='Target_Beneficiary',right_on='BeneID').drop(columns=["BeneID"])

## exporting the dataset
all_recommendation.to_csv("../data/processed_data/all_recommendations_0.8_top5.csv",index=False)

# exporting the trained model to a file using pickle
import pickle
with open('../models/Recommendations_model.pkl', 'wb') as f:
    pickle.dump(recommendation_model, f)

all_recommendation.head(25)

len(all_recommendation)
